Puppy Lover Page
Welcome to the Puppy Lover Page repository! 🐾

This project is a simple HTML and CSS webpage dedicated to celebrating our love for puppies. Explore five adorable puppy pictures showcased in three engaging sections. The page features a header, footer, and an H1 title to enhance navigation and readability.

Immerse yourself in the heartwarming world of puppies and indulge in their cuteness. Whether you're a devoted puppy enthusiast or just in need of a smile, this page is the perfect destination.

Join us in spreading joy and happiness through the charm of our furry friends. Feel free to browse the adorable pictures, share your favorite puppy moments, and contribute to making this page even more delightful!

Happy puppy browsing! 🐶💖